% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{article}
\alias{article}
\title{Function for article template}
\usage{
article(other_css = NULL, ...)
}
\arguments{
\item{other_css}{Add an other CSS}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
a pagedown template
}
\description{
Function for article template
}
